Run test with pdo_mysql & mysqli
