+++ features/bootstrap/FeatureContext.php
@@ -1,16 +1,11 @@
 <?php
 
-use Behat\Behat\Context\Context;
-use Behat\Behat\Context\SnippetAcceptingContext;
 use Behat\Gherkin\Node\PyStringNode;
 use Behat\Gherkin\Node\TableNode;
 use Behat\Behat\Tester\Exception\PendingException;
 use Doctrine\DBAL\DriverManager;
 
-/**
- * Defines application features from the specific context.
- */
-class FeatureContext implements Context, SnippetAcceptingContext
+trait FeatureContext
 {
     private $connections = [];
 
@@ -341,22 +336,6 @@ SQL;
             );
         }
         return $this->connections[$connectionName]['instance'];
-    }
-
-    private function masterParams()
-    {
-        return [
-            'driver'        => 'mysqli',
-            'host'          => getenv('DBLINKER_DB_1_PORT_3306_TCP_ADDR'),
-            'user'          => getenv('DBLINKER_DB_1_ENV_MYSQL_USER'),
-            'password'      => getenv('DBLINKER_DB_1_ENV_MYSQL_PASSWORD'),
-            'dbname'        => getenv('DBLINKER_DB_1_ENV_MYSQL_DATABASE'),
-            'driver'        => 'pdo_mysql',
-            'driverOptions' => [
-                // todo move to retry driver
-                PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION,
-            ],
-        ];
     }
 }
